requirements.txt
langgraph>=0.0.50
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0